pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]

[project]
name = "imagedgd"
description = "Deep Gaussian Decoder for Image Generation with MLOps"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "scikit-learn>=1.1.0",
    "umap-learn>=0.5.3",
    "mlflow>=2.8.0",
    "optuna>=3.4.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "click>=8.1.0",
    "tqdm>=4.64.0",
    "pandas>=1.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.11.0",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
    "tensorboard>=2.11.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "black>=22.10.0",
    "flake8>=5.0.0",
    "isort>=5.10.0",
    "pre-commit>=2.20.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.16.0",
]

[project.scripts]
imagedgd = "scripts.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.setuptools.package-dir]
"" = "."
"imagedgd" = "src"

[tool.setuptools_scm]
write_to = "src/_version.py"
